Fix race condition on reconnect
This commit removes the unnecessary and problematic call to reconnect.
The underlying EventMachine::Hiredis::Client already handle the reconnection every 0.5 seconds. Having a second timer also running every 0.5s can make the two reconnections overlapse and eventually it can make the reconnection to fail for a while.
The reason is that evenmachine C extension uses non blocking sockets so if the two timers are executed during the same event loop the execution will be as follow:
- Execute timer with reconnect call (from em-hiredis) -> this will open a socket and will not wait for it to be oppened
- Execute timer with reconnect! call (from faye-redis-ruby) -> this will close the socket open from the previous timer and call a reconnect at next_tick
- next_tick is executed:
                         - All sockets are read from and/or written unless for sockets marked as closed as in our case.
                         - Reconnect is executed but notices we already have one socket open for this redis connection so it does nothing.
- _CleanupSockets is called so our redis socket is closed which is triggering the reconnect_failed event and this can goes on for a while until the 2 timers are not executed during the same event loop.

The code for eventmachine can be seen in the em.cpp file in eventmachine github repository.