Fail fast when redis is not accessible without sending tcp FIN RD-25611
This commit handles scenario where HA is handled via DNS update as this is down in elasticache.
In this scenario if redis master crashes or if it's not accessible anymore elasticache will promote a secondary as master and update the DNS to point to the new master.

However the redis connection on client side has no way to know it should close the socket since it didn't receive a FIN or RST packet.
The way tcp handles this is to retransmit packet up to 15 times (/proc/sys/net/ipv4/tcp_retries2) with a backoff strategy of **2 capped to 60 seconds.
This means it will timeout after ~ 8 minutes or when the old primary is back up to send a RST which is also takes about 8 minutes on AWS.

This commit takes advantage of the implementation done in em-hiredis that exposes a periodic timer to check connections. This commits exposes it to end user.